+++ CHANGES.rst
@@ -10,6 +10,8 @@ Features
 - Issue #76: Extend testing helpers to allow customization of the Java
   classpath by specifying the new `ZOOKEEPER_CLASSPATH` environment variable.
 
+- Issue #65: Allow non-blocking semaphore acquisition.
+
 Bug Handling
 ************